requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2>=2.7,<3.0
mysqlclient